Support parameterized cache size of SGX PFS (apply intel's patch)

+++ sdk/protected_fs/sgx_tprotected_fs/lru_cache.h
@@ -83,7 +83,7 @@ public:
 	lru_cache();
 	~lru_cache();
 
-	void rehash(uint32_t size_);
+	void rehash(uint64_t size_);
 
 	bool add(uint64_t key, void* p);
 	void* get(uint64_t key);
